requirements.txt
yt-dlp>=2025.1.1
faster-whisper>=1.0.0
opencv-python-headless>=4.10.0.84
imagehash>=4.3.1
Pillow>=10.4.0
scikit-image>=0.24.0
pytesseract